Add files via upload
This is my first GUI project. I learned this from Youtube to get some experience with animation. This is written in Python. This is two balls bouncing in a animation board. I will be learning more about animation soon. This is me getting my foot in the door for animation.